Tagged every non-Biblical quote in WP (first pass)
Also restored tokens that were lost in conversion on fol. 194 because the converter didn't know what to do with TITUS's way of indicating lacunae.

I tried to tag the actual quoted/paraphrased text, excluding the attribution ("As Athanasius writes in his" etc.). Where there was only an attribution without a quote, I tagged the attribution so that it can be found and re-tagged appropriately later.